requirements.txt
Django==2.0.5
django-extensions==2.0.7
django-filter==1.1.0
django-hashid-field==2.1.1
django-picklefield==1.0.0
djangorestframework==3.8.2
factory-boy==2.11.1
hashids==1.2.0
Markdown==2.6.11
mysqlclient==1.3.12
pytz==2018.4
six==1.11.0